Added remove dead token code to updateTokens API as well.
