+++ coin-alert/src/app/api/checkPriceAlerts/route.ts
@@ -2,9 +2,7 @@ import { AlarmConfig } from "@/app/lib/constants/alarmConstants";
 import { getAllUsers, RecentNotification, SirenUser } from "@/app/lib/firebase/userUtils";
 import { calculatePriceChange, getAlarmConfig, getLastHourPrices, getTokensFromBlockchain, NotificationReturn } from "@/app/lib/utils/priceAlertHelper";
 import { sendNotification } from "../../lib/sendNotifications"; // Push notification logic
-import { getTokenCached, Token } from "@/app/lib/firebase/tokenUtils";
-import { deleteDoc, doc } from "firebase/firestore";
-import { db } from "@/app/lib/firebase/firebase";
+import { getTokenCached, removeTokenIfDead, Token } from "@/app/lib/firebase/tokenUtils";
 
 
 let tokensCache: Map<string, Token> = new Map<string, Token>()
@@ -44,41 +42,6 @@ function isTokenMinuteAfterCooldown(
   const elapsedTime = (now - lastNotificationTime) / (60 * 1000); // Convert to minutes
 
   return elapsedTime > minutes; // ✅ Return true if notification is older than cooldown
-}
-
-
-async function removeTokenIfDead(token: string, tokenDb: Token | undefined): Promise<boolean> {
-  try {
-    if (!tokenDb || !tokenDb.prices || tokenDb.prices.length < 2) {
-      console.log(`🔹 Not enough price data to determine if ${token} is dead.`);
-      return false;
-    }
-
-    const prices = tokenDb.prices;
-    const oneHourMs = 60 * 60 * 1000;
-
-    // 🔹 Check if any two price entries are more than 60 minutes apart and identical
-    for (let i = 0; i < prices.length - 1; i++) {
-      for (let j = i + 1; j < prices.length; j++) {
-        const timeDifference = Math.abs(prices[j].timestamp - prices[i].timestamp);
-        if (timeDifference > oneHourMs && prices[i].price === prices[j].price) {
-          console.log(`💀 Token ${token} detected as dead (same price > 60 min apart). Removing...`);
-
-          // 🔹 Remove the token from Firestore
-          const tokenDocRef = doc(db, "uniqueTokens", token);
-          await deleteDoc(tokenDocRef);
-          console.log(`✅ Token ${token} successfully removed from Firestore.`);
-          return true;
-        }
-      }
-    }
-
-    console.log(`🔹 Token ${token} is still active.`);
-    return false;
-  } catch (error) {
-    console.error(`❌ Error removing dead token ${token}:`, error);
-    return false;
-  }
 }
