+++ coin-alert/src/app/lib/firebase/tokenUtils.ts
@@ -1,4 +1,4 @@
-import { doc, FirestoreDataConverter, getDoc, Timestamp } from "firebase/firestore";
+import { deleteDoc, doc, FirestoreDataConverter, getDoc, Timestamp } from "firebase/firestore";
 import { db } from "../firebase/firebase";
 
 export interface PriceData {
@@ -47,6 +47,41 @@ export async function getTokenCached(token: string, tokenCache: Map<string, Toke
     return tokenDb
   }
 
+}
+
+
+export async function removeTokenIfDead(token: string, tokenDb: Token | undefined): Promise<boolean> {
+  try {
+    if (!tokenDb || !tokenDb.prices || tokenDb.prices.length < 2) {
+      console.log(`🔹 Not enough price data to determine if ${token} is dead.`);
+      return false;
+    }
+
+    const prices = tokenDb.prices;
+    const oneHourMs = 60 * 60 * 1000;
+
+    // 🔹 Check if any two price entries are more than 60 minutes apart and identical
+    for (let i = 0; i < prices.length - 1; i++) {
+      for (let j = i + 1; j < prices.length; j++) {
+        const timeDifference = Math.abs(prices[j].timestamp - prices[i].timestamp);
+        if (timeDifference > oneHourMs && prices[i].price === prices[j].price) {
+          console.log(`💀 Token ${token} detected as dead (same price > 60 min apart). Removing...`);
+
+          // 🔹 Remove the token from Firestore
+          const tokenDocRef = doc(db, "uniqueTokens", token);
+          await deleteDoc(tokenDocRef);
+          console.log(`✅ Token ${token} successfully removed from Firestore.`);
+          return true;
+        }
+      }
+    }
+
+    console.log(`🔹 Token ${token} is still active.`);
+    return false;
+  } catch (error) {
+    console.error(`❌ Error removing dead token ${token}:`, error);
+    return false;
+  }
 }
 
 export const tokenConverter: FirestoreDataConverter<Token> = {
